requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
Flask==2.0.3
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Naked==0.1.31
passlib==1.7.4
pycrypto==2.6.1
Python-dev==2.0.0.dev0
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
zipp==3.6.0
